fix(page): Move sidebar interaction to “Page” layout
This change happens to avoid too many dependencies on `sidebarFn` prop to close and open the sidebar. Keeping this way the layout simple and effective.

TBD: It will also be an improvement for accessibility when changing pages to move back users to the top of the content on page load and with the sidebar closed.