feat(page): Create new component (#4)
* feat(content): Create new container

* feat(page): Create new layout

* fix: Update pages to use “Page” component

* fix(page): Make content full height on large screens

* fix: Make sidebar interactive in small screens

- Create a function to open the sidebar.
- Create a function to close the sidebar.
- Change the sidebar state based on resize event.

* fix(sidebar): Keep border for all screens

When the sidebar is below the breakpoint, it hides the shadow that acts as a border. This is a problem because visually the sidebar gets lost when the content is full width, because both (sidebar and content) have white background.

* fix(page): Move sidebar interaction to “Page” layout

This change happens to avoid too many dependencies on `sidebarFn` prop to close and open the sidebar. Keeping this way the layout simple and effective.

TBD: It will also be an improvement for accessibility when changing pages to move back users to the top of the content on page load and with the sidebar closed.